+++ aiohomematic/model/custom/switch.py
@@ -9,7 +9,6 @@ Public API of this module is defined by __all__.
 from __future__ import annotations
 
 from collections.abc import Mapping
-from enum import StrEnum
 import logging
 from typing import Any, Final
 
@@ -17,6 +16,7 @@ from aiohomematic.const import DataPointCategory, DeviceProfile, Field, Paramete
 from aiohomematic.interfaces import ChannelProtocol
 from aiohomematic.model.custom import definition as hmed
 from aiohomematic.model.custom.data_point import CustomDataPoint
+from aiohomematic.model.custom.mixins import GroupStateMixin, StateChangeTimerMixin
 from aiohomematic.model.custom.support import CustomConfig, ExtendedConfig
 from aiohomematic.model.data_point import CallParameterCollector, bind_collector
 from aiohomematic.model.generic import DpAction, DpBinarySensor, DpSwitch
@@ -25,14 +25,7 @@ from aiohomematic.property_decorators import state_property
 _LOGGER: Final = logging.getLogger(__name__)
 
 
-class _StateChangeArg(StrEnum):
-    """Enum with switch state change arguments."""
-
-    OFF = "off"
-    ON = "on"
-
-
-class CustomDpSwitch(CustomDataPoint):
+class CustomDpSwitch(StateChangeTimerMixin, GroupStateMixin, CustomDataPoint):
     """Class for Homematic switch data point."""
 
     __slots__ = (
@@ -43,11 +36,6 @@ class CustomDpSwitch(CustomDataPoint):
 
     _category = DataPointCategory.SWITCH
 
-    @property
-    def group_value(self) -> bool | None:
-        """Return the current group value of the switch."""
-        return self._dp_group_state.value
-
     @state_property
     def value(self) -> bool | None:
         """Return the current channel value of the switch."""
@@ -55,13 +43,7 @@ class CustomDpSwitch(CustomDataPoint):
 
     def is_state_change(self, **kwargs: Any) -> bool:
         """Check if the state changes due to kwargs."""
-        if (on_time_running := self.timer_on_time_running) is not None and on_time_running is True:
-            return True
-        if self.timer_on_time is not None:
-            return True
-        if kwargs.get(_StateChangeArg.ON) is not None and self.value is not True:
-            return True
-        if kwargs.get(_StateChangeArg.OFF) is not None and self.value is not False:
+        if self.is_state_change_for_on_off(**kwargs):
             return True
         return super().is_state_change(**kwargs)